requirements.txt
# HTTP客户端库，用于发送API请求
requests==2.31.0

# 异步HTTP服务器框架
aiohttp==3.9.1
aiohttp-cors==0.7.0

# 环境变量管理
python-dotenv==1.0.0

# 智谱AI SDK
zhipuai==1.0.7 